// web/src/__generated__/schemas/OfferEntitlementsResponse.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { OfferEntitlement } from './OfferEntitlement';

export type OfferEntitlementsResponse = { 
  entitlements?: OfferEntitlement[]; 
  playerId?: string; 
};