// web/src/__generated__/schemas/RedeemOfferRequest.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { OfferRedeemPayload } from './OfferRedeemPayload';

export type RedeemOfferRequest = { 
  federationId: string; 
  grantId: string; 
  playerId: string; 
  request?: OfferRedeemPayload; 
};